starting to change play function

+++ velhia.game/src/entities/match/match.py
@@ -4,12 +4,13 @@ from .player import Player
 from .play import Play
 from src.shared.types.char import Char
 
+
 class Match(TypedDict):
 
     _id: str
     begin: str
     end: str
-    time: int
+    time: float
     sa: Player
     mas: MAS
     plays: List[Play]
